Remove potential panic
You are logging the error here but still calling the returned `handler` regardless of the value of `handler`, which can be nil.

Additionally, calling `log.Fatal()` won't call deferred functions. You are better off changing `StartServer` to return an error and return an error instead of `log.Fatal()`. In the for-loop though, you can log and then use `continue` so you log, but just move on to the next request. 

Thanks!